All refactoring done, and working
finally done
